Unify DrawFramebufferToOutput
